Fail to start on MSSQL with delete_domain_limit

+++ src/mam/mod_mam_rdbms_arch.erl
@@ -165,28 +165,28 @@ prepare_remove_domain(#{delete_domain_limit := infinity}) ->
     mongoose_rdbms:prepare(mam_remove_domain_users, mam_server_user, [server],
                            <<"DELETE FROM mam_server_user WHERE server = ?">>);
 prepare_remove_domain(#{delete_domain_limit := Limit}) ->
-    {MaybeLimitSQL, MaybeLimitMSSQL} = rdbms_queries:get_db_specific_limits_binaries(Limit),
-    IdTable = <<"(SELECT * FROM (SELECT ", MaybeLimitMSSQL/binary,
-                                " msg.user_id, msg.id FROM mam_message msg",
-                                " INNER JOIN mam_server_user msu ON msu.id=msg.user_id",
-                                " WHERE msu.server = ? ", MaybeLimitSQL/binary, ") AS T)">>,
+    LimitSQL = case mongoose_rdbms:db_type() of
+                        mssql -> throw(delete_domain_limit_not_supported_for_mssql);
+                        _ -> {MaybeLimitSQL, _} = rdbms_queries:get_db_specific_limits_binaries(Limit),
+                             MaybeLimitSQL
+                    end,
+    IdTable = <<"(SELECT * FROM ",
+                    "(SELECT msg.user_id, msg.id FROM mam_message msg",
+                    " INNER JOIN mam_server_user msu ON msu.id=msg.user_id",
+                    " WHERE msu.server = ? ", LimitSQL/binary, ") AS T)">>,
     mongoose_rdbms:prepare(mam_incr_remove_domain, mam_message, ['mam_server_user.server'],
-                           <<"DELETE FROM mam_message "
-                             "WHERE (user_id, id) IN ", IdTable/binary>>),
-
-    CfgTable = <<"(SELECT * FROM (SELECT ", MaybeLimitMSSQL/binary,
-                                " cfg.user_id, cfg.remote_jid FROM mam_config cfg",
-                                " INNER JOIN mam_server_user msu ON msu.id=cfg.user_id",
-                                " WHERE msu.server = ? ", MaybeLimitSQL/binary, ") AS T)">>,
+                           <<"DELETE FROM mam_message WHERE (user_id, id) IN ", IdTable/binary>>),
+    CfgTable = <<"(SELECT * FROM ",
+                     "(SELECT cfg.user_id, cfg.remote_jid FROM mam_config cfg",
+                     " INNER JOIN mam_server_user msu ON msu.id=cfg.user_id",
+                     " WHERE msu.server = ? ", LimitSQL/binary, ") AS T)">>,
     mongoose_rdbms:prepare(mam_incr_remove_domain_prefs, mam_config, ['mam_server_user.server'],
                            <<"DELETE FROM mam_config "
                              "WHERE (user_id, remote_jid) IN ", CfgTable/binary>>),
-
-    ServerTable = <<"(SELECT * FROM (SELECT ", MaybeLimitMSSQL/binary,
-                                    " id FROM mam_server_user WHERE server = ? ", MaybeLimitSQL/binary, ") as t)">>,
+    ServerTable = <<"(SELECT * FROM ",
+                    "(SELECT id FROM mam_server_user WHERE server = ? ", LimitSQL/binary, ") as t)">>,
     mongoose_rdbms:prepare(mam_incr_remove_domain_users, mam_server_user, [server],
-                           <<"DELETE ", MaybeLimitMSSQL/binary,
-                             " FROM mam_server_user WHERE id IN ", ServerTable/binary>>).
+                           <<"DELETE FROM mam_server_user WHERE id IN ", ServerTable/binary>>).
 
 
 %% ----------------------------------------------------------------------
